src/werewolf/requirements.txt
Django==1.5.2
django-reversion==1.7.1
South==0.8.2

# Werewolf old dependences
# 0.1
#Django==1.5.1
#django-reversion==1.6.5
